HiddenTubeController
script that is meant to check whether the two conditions required to be met to open the final door in room 3 are fulfilled.

not particularly ideal as instead of checking every frame, it would be more efficient to check after every time a condition is met, as this way it will check twice (one for each required condition) and pass on the second check, but as there were no puzzles to currently reference, this was the best i could do. will refine in the future.